+++ database/migrations/2020_11_14_123812_create_alipay_table.php
@@ -15,7 +15,7 @@ class CreateAlipayTable extends Migration
     {
         Schema::create('alipay', function (Blueprint $table) {
             $table->tinyIncrements('id');
-            $table->unsignedTinyInteger('app_id')->comment('应用id');
+            $table->unsignedSmallInteger('app_id')->comment('应用id');
             $table->string('open_id')->comment('支付宝的app_id');
             $table->text('ali_public_key')->comment('支付宝公钥，用于验签支付宝回调数据');
             $table->text('private_key')->comment('应用私钥，用于下单生成签名');
