+++ database/migrations/2020_11_14_141055_create_wechat_table.php
@@ -15,7 +15,7 @@ class CreateWechatTable extends Migration
     {
         Schema::create('wechat', function (Blueprint $table) {
             $table->tinyIncrements('id');
-            $table->unsignedTinyInteger('app_id')->comment('应用id');
+            $table->unsignedSmallInteger('app_id')->comment('应用id');
             $table->string('open_id')->comment('微信的app_id');
             $table->string('mch_id');
             $table->string('app_secret');
